users_service/requirements.txt
asyncpg==0.25.0
fastapi==0.77.1
pydantic==1.9.0
python-dotenv==0.20.0
uvicorn==0.17.6
sqlmodel==0.0.6
Faker==13.11.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
SQLAlchemy==1.4.36
python-multipart==0.0.5